fix(web): 轻微问题修复 (#21 #22 + Base64 统一 + type 安全)
- #21 || 改 ??（sessionStore, ItemDetailPage）
- #22 SendCodeButton 组件卸载清理 interval
- Base64: 删 pbkdf2.ts 重复定义，LoginPage 改用 base64ToBytes 导入
- GOOGLE_CLIENT_ID 支持 VITE_ 环境变量覆盖
- catch (e: any) 改 catch (e) + instanceof Error
- password.ts 随机密码 Uint8Array 改 Uint32Array 减少模偏差

+++ web/src/utils/password.ts
@@ -30,7 +30,8 @@ export function generatePassword(opts?: Partial<PasswordOptions>): string {
   if (options.includeSpecials) pool += SPECIALS;
   if (pool.length === 0) pool = LOWERCASE + DIGITS;
 
-  const chars = new Uint8Array(options.length);
+  // 用 Uint32Array + 取模 避免 Uint8Array % pool.length 的微小模偏差
+  const chars = new Uint32Array(options.length);
   crypto.getRandomValues(chars);
   let result = "";
   for (let i = 0; i < options.length; i++) {
